Adding generic llm converter
